1.api接口问题修改 2.收藏排序问题 3.导出

+++ netInsight-alert/src/main/java/com/trs/netInsight/widget/analysis/service/IChartAnalyzeService.java
@@ -566,7 +566,7 @@ public interface IChartAnalyzeService {
 	 * @return
 	 */
 	public List<Map<String, String>> emotionOption(QueryBuilder searchBuilder, SpecialProject specialProject);
-	public int getSituationAssessment(QueryBuilder searchBuilder, SpecialProject specialProject) throws TRSException;
+	public double getSituationAssessment(QueryBuilder searchBuilder, SpecialProject specialProject) throws TRSException;
 
 	/**
 	 * 获取专题内图表列表数据
